moved ftp-config to ftpconfig.xml

+++ ftp.sh
@@ -1,25 +1,33 @@
 #!/bin/bash
-DIR="/home/USER/weather"
-SFTPPASS="PASS OR REPLACE WITH KEY"
-SFTPUSER="USER@DOMAIN.TLD"
-URL="DOMAIN.TLD/PATH/FILE.EXT"
+CONFIG="ftpconfig.xml"
+FTP_COMMAND="mput temp*
+mput humi*
+mput ambi*
+mput baro*
+quit
+"
 
-cd $DIR
+if [ ! -f $CONFIG ]; then
+	echo "configuration file not found, exiting! (see ftpconfig.sample.xml)"
+	exit 1;
+fi
+read DIR SFTPUSER URL SFTPPASS < <(xmlstarlet sel -t -v "//dir" -o $'\t' -v "//sftpuser" -o $'\t' -v "//url" -o $'\t' -v "//sftppass" $CONFIG)
+
+cd $DIR || exit 1
 cp records/humi* ftp/
 cp records/ambi* ftp/
 cp records/temp* ftp/
 cp records/baro* ftp/
 cd ftp
-# Using a plain password here is evil, but in my context sadly needed
-# better use key-authentication, and switch the commented lines
-#sftp -oBatchMode=no -b - $SFTPUSER <<EOF
-sshpass -p $SFTPPASS sftp -oBatchMode=no -b - $SFTPUSER <<EOF
-mput temp*
-mput humi*
-mput ambi*
-mput baro*
-quit
-EOF
+
+
+if [ -z $SFTPPASS ]; then
+	#sftp-key-auth
+	sftp -oBatchMode=no -b - $SFTPUSER "$FTP_COMMAND"
+else
+	#sftp-pass-auth
+	sshpass -p $SFTPPASS sftp -oBatchMode=no -b - $SFTPUSER "$FTP_COMMAND"
+fi
 rm humi* ambi* temp* baro*
 cd ..
 wget $URL -O logs/wget_recent -q
